Slight changes
We will now only synchronize the modules once, ie. reset clocks. This is when we reboot the modules. If we synchronize the modules each time we start a run then the clock will have a random offset between each other, thus we will need different alignment of the time spectra for each run. This will cause a lot of extra work that can easily be avoided!