New features
I have added new features. Now it is possible to save settings As, the program will save settings every time it writes to modules and the settings will be backed up everytime we try to write to disk.
This version also allows the user to copy settings of one channel to several other channels, making it much easier to change settings. I have also changed a bunch of small tings so that we now will get much less warnings while compiling.