Added Baseline average field
I have added Baseline Average field and changed such that the UI will actually update the baseline values. Also I have added an option to read the baseline cut of only a single channel at a time. The baseline offset is no longer fixed when the run starts, but by the user in the UI.