Set filename to whatever the current file is
This should ensure that whatever the current file is, we will be writing the scalers with the same name. That means that we will write:
sirius-YYYYMMDD-HHMMSS.data and sirius-YYYYMMDD-HHMMSS.data.stats
sirius-YYYYMMDD-HHMMSS-big-000.data and sirius-YYYYMMDD-HHMMSS.data.stats
etc.

The goal is to ensure that we never lose the scalers, even if the DAQ crashes.