Minor changes
I've made a minor change that if the filename is set to empty then we will write to the /dev/null. That way we do not need to think about if tests later to check if the pointer is set or not.
I've also moved the XIA_end_run to be closer to the StartLMR as it is more natural to keep these two close together.